+++ src/commands/deployWorkspaceProject/internal/DeployWorkspaceProjectConfirmStep.ts
@@ -4,10 +4,10 @@
  *--------------------------------------------------------------------------------------------*/
 
 import { nonNullValue } from "@microsoft/vscode-azext-utils";
-import { ext } from "../../extensionVariables";
-import { localize } from "../../utils/localize";
-import { OverwriteConfirmStepBase } from "../OverwriteConfirmStepBase";
-import { type DeployWorkspaceProjectContext } from "./DeployWorkspaceProjectContext";
+import { ext } from "../../../extensionVariables";
+import { localize } from "../../../utils/localize";
+import { OverwriteConfirmStepBase } from "../../OverwriteConfirmStepBase";
+import { type DeployWorkspaceProjectContext } from "../DeployWorkspaceProjectContext";
 
 export class DeployWorkspaceProjectConfirmStep extends OverwriteConfirmStepBase<DeployWorkspaceProjectContext> {
     constructor(private readonly suppressConfirmation: boolean) {
