+++ src/commands/deployWorkspaceProject/internal/deployWorkspaceProjectInternal.ts
@@ -6,26 +6,26 @@
 import { LocationListStep, ResourceGroupCreateStep } from "@microsoft/vscode-azext-azureutils";
 import { AzureWizard, GenericTreeItem, activityInfoIcon, activitySuccessContext, nonNullValueAndProp, type AzureWizardExecuteStep, type AzureWizardPromptStep, type ExecuteActivityContext } from "@microsoft/vscode-azext-utils";
 import { ProgressLocation, window } from "vscode";
-import { appProvider, managedEnvironmentsId } from "../../constants";
-import { ext } from "../../extensionVariables";
-import { type ContainerAppItem } from "../../tree/ContainerAppItem";
-import { type ManagedEnvironmentItem } from "../../tree/ManagedEnvironmentItem";
-import { createActivityChildContext, createActivityContext } from "../../utils/activity/activityUtils";
-import { getVerifyProvidersStep } from "../../utils/getVerifyProvidersStep";
-import { localize } from "../../utils/localize";
-import { type IContainerAppContext } from "../IContainerAppContext";
-import { ContainerAppCreateStep } from "../createContainerApp/ContainerAppCreateStep";
-import { LogAnalyticsCreateStep } from "../createManagedEnvironment/LogAnalyticsCreateStep";
-import { ManagedEnvironmentCreateStep } from "../createManagedEnvironment/ManagedEnvironmentCreateStep";
-import { ContainerAppUpdateStep } from "../image/imageSource/ContainerAppUpdateStep";
-import { ImageSourceListStep } from "../image/imageSource/ImageSourceListStep";
-import { IngressPromptStep } from "../ingress/IngressPromptStep";
+import { appProvider, managedEnvironmentsId } from "../../../constants";
+import { ext } from "../../../extensionVariables";
+import { type ContainerAppItem } from "../../../tree/ContainerAppItem";
+import { type ManagedEnvironmentItem } from "../../../tree/ManagedEnvironmentItem";
+import { createActivityChildContext, createActivityContext } from "../../../utils/activity/activityUtils";
+import { getVerifyProvidersStep } from "../../../utils/getVerifyProvidersStep";
+import { localize } from "../../../utils/localize";
+import { type IContainerAppContext } from "../../IContainerAppContext";
+import { ContainerAppCreateStep } from "../../createContainerApp/ContainerAppCreateStep";
+import { LogAnalyticsCreateStep } from "../../createManagedEnvironment/LogAnalyticsCreateStep";
+import { ManagedEnvironmentCreateStep } from "../../createManagedEnvironment/ManagedEnvironmentCreateStep";
+import { ContainerAppUpdateStep } from "../../image/imageSource/ContainerAppUpdateStep";
+import { ImageSourceListStep } from "../../image/imageSource/ImageSourceListStep";
+import { IngressPromptStep } from "../../ingress/IngressPromptStep";
+import { type DeployWorkspaceProjectContext } from "../DeployWorkspaceProjectContext";
+import { DefaultResourcesNameStep } from "../getDefaultValues/DefaultResourcesNameStep";
+import { getDefaultContextValues } from "../getDefaultValues/getDefaultContextValues";
 import { DeployWorkspaceProjectConfirmStep } from "./DeployWorkspaceProjectConfirmStep";
-import { type DeployWorkspaceProjectContext } from "./DeployWorkspaceProjectContext";
 import { DeployWorkspaceProjectSaveSettingsStep } from "./DeployWorkspaceProjectSaveSettingsStep";
 import { ShouldSaveDeploySettingsPromptStep } from "./ShouldSaveDeploySettingsPromptStep";
-import { DefaultResourcesNameStep } from "./getDefaultValues/DefaultResourcesNameStep";
-import { getDefaultContextValues } from "./getDefaultValues/getDefaultContextValues";
 
 export type DeployWorkspaceProjectInternalContext = IContainerAppContext & Partial<DeployWorkspaceProjectContext>;
