+++ src/commands/deployWorkspaceProject/settings/DeployWorkspaceProjectSettingsV1.ts
@@ -1,0 +1,10 @@
+/*---------------------------------------------------------------------------------------------
+ *  Copyright (c) Microsoft Corporation. All rights reserved.
+ *  Licensed under the MIT License. See License.txt in the project root for license information.
+ *--------------------------------------------------------------------------------------------*/
+
+export interface DeployWorkspaceProjectSettingsV1 {
+    containerAppResourceGroupName?: string;
+    containerAppName?: string;
+    containerRegistryName?: string;
+}
